FindBugs fixes
git-svn-id: file:///home/svn/omero/trunk@4319 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/server/src/ome/logic/RepositoryInfoImpl.java
@@ -324,7 +324,7 @@ public class RepositoryInfoImpl extends AbstractLevel2Service implements
         }
 
         // cleanup any pixels
-        if (pixels != null & pixels.size() > 0) {
+        if (pixels != null && pixels.size() > 0) {
             log.info("Removing pixels: " + pixels);
             pixelsService.removePixels(pixels);
         }
